adding counter

+++ Models/Search/SearchHit.cs
@@ -4,11 +4,10 @@
 	{
 		public string Title { get; set; }
 
-		public double Price { get; set; }
+		public decimal Price { get; set; }
 
 		public string Image { get; set; }
 
 		public string Description { get; set; }
-
 	}
 }
